Hodor quasi terminé, seul problème est que Debian et Wall ne s'entendent pas bien

+++ hodor.sh
@@ -1,5 +1,9 @@
-
-echo "HODOOOR !!!" | wall
-E="$(cat /dev/urandom | tr -dc '[:alpha:]' | head -c10)"
-sudo tar -zcvf /var/log/exercice/archives/$E.tar.gz /var/log/exercice/recent_logs
-sudo rm /var/log/exercice/recent_logs/*
+A="$(df -i / | cut -c 40-42 | cut -d ' ' -f 2 | tail -1)"
+if [ $A -ge 12 ]; then
+	echo 'HODOOOR!!!'
+	B="$(mktemp)"
+	E="$(cat /dev/urandom | tr -dc '[:alpha:]' | head -c10)"
+	sudo tar -zcvf /var/log/exercice/archives/$E.tar.gz /var/log/exercice/recent_logs > $B
+	sudo rm /var/log/exercice/recent_logs/*
+	rm $B
+fi
